requirements.txt
streamlit==1.51.0
langchain==0.3.4
langchain-community==0.3.3
langchain-openai==0.2.2
chromadb==0.5.11
tiktoken==0.8.0
pypdf==5.1.0
python-dotenv==1.0.1